+++ src/main/kotlin/com/cognifide/gradle/aem/AemExtension.kt
@@ -15,7 +15,7 @@ import com.cognifide.gradle.aem.common.pkg.vault.FilterFile
 import com.cognifide.gradle.aem.instance.*
 import com.cognifide.gradle.aem.pkg.PackagePlugin
 import com.cognifide.gradle.aem.pkg.tasks.PackageCompose
-import com.cognifide.gradle.aem.instance.rcp.RcpClient
+import com.cognifide.gradle.aem.common.instance.rcp.RcpClient
 import com.cognifide.gradle.aem.common.pkg.vault.VaultClient
 import com.cognifide.gradle.aem.common.pkg.vault.VaultSummary
 import com.cognifide.gradle.aem.pkg.PackageSyncPlugin
