+++ src/main/kotlin/com/cognifide/gradle/aem/common/instance/AbstractInstance.kt
@@ -66,6 +66,10 @@ abstract class AbstractInstance(@Transient @JsonIgnore protected val aem: AemExt
 
     fun provision() = manager.provisioner.provision(this)
 
+    fun satisfy() = manager.satisfier.satisfy(this)
+
+    fun tail() = manager.tailer.tail(this)
+
     override fun equals(other: Any?): Boolean {
         if (this === other) return true
         if (javaClass != other?.javaClass) return false
