+++ src/main/kotlin/com/cognifide/gradle/aem/common/instance/rcp/RcpSummary.kt
@@ -1,4 +1,4 @@
-package com.cognifide.gradle.aem.instance.rcp
+package com.cognifide.gradle.aem.common.instance.rcp
 
 import com.cognifide.gradle.aem.common.instance.Instance
 import com.cognifide.gradle.common.utils.Formats
