+++ src/main/kotlin/com/cognifide/gradle/aem/common/instance/tail/LogChunk.kt
@@ -1,9 +1,8 @@
-package com.cognifide.gradle.aem.instance.tail
+package com.cognifide.gradle.aem.common.instance.tail
 
 import com.cognifide.gradle.aem.common.instance.Instance
 
 class LogChunk(val instance: Instance, val logs: List<Log>) {
 
-    val size: Int
-        get() = logs.size
+    val size: Int get() = logs.size
 }
