+++ src/main/kotlin/com/cognifide/gradle/aem/common/instance/tail/LogDestination.kt
@@ -1,4 +1,4 @@
-package com.cognifide.gradle.aem.instance.tail
+package com.cognifide.gradle.aem.common.instance.tail
 
 interface LogDestination {
     fun dump(logs: List<Log>)
