+++ src/main/kotlin/com/cognifide/gradle/aem/common/instance/tail/io/UrlSource.kt
@@ -1,8 +1,8 @@
-package com.cognifide.gradle.aem.instance.tail.io
+package com.cognifide.gradle.aem.common.instance.tail.io
 
 import com.cognifide.gradle.aem.common.instance.Instance
-import com.cognifide.gradle.aem.instance.tail.Tailer
-import com.cognifide.gradle.aem.instance.tail.LogSource
+import com.cognifide.gradle.aem.common.instance.tail.Tailer
+import com.cognifide.gradle.aem.common.instance.tail.LogSource
 import com.cognifide.gradle.common.http.RequestException
 import java.io.BufferedReader
