+++ src/main/kotlin/com/cognifide/gradle/aem/instance/LocalInstancePlugin.kt
@@ -1,7 +1,7 @@
 package com.cognifide.gradle.aem.instance
 
 import com.cognifide.gradle.aem.instance.tasks.InstanceProvision
-import com.cognifide.gradle.aem.instance.rcp.InstanceRcp
+import com.cognifide.gradle.aem.instance.tasks.InstanceRcp
 import com.cognifide.gradle.aem.instance.tasks.InstanceSatisfy
 import com.cognifide.gradle.aem.instance.tasks.InstanceTail
 import com.cognifide.gradle.aem.instance.tasks.*
