--- src/main/kotlin/com/cognifide/gradle/aem/instance/tail/InstanceTailerException.kt
@@ -1,5 +1,0 @@
-package com.cognifide.gradle.aem.instance.tail
-
-import com.cognifide.gradle.aem.AemException
-
-class InstanceTailerException(message: String) : AemException(message)
