+++ src/test/kotlin/com/cognifide/gradle/aem/instance/InstancePluginTest.kt
@@ -2,7 +2,7 @@ package com.cognifide.gradle.aem.instance
 
 import com.cognifide.gradle.aem.AemExtension
 import com.cognifide.gradle.aem.common.utils.using
-import com.cognifide.gradle.aem.instance.rcp.InstanceRcp
+import com.cognifide.gradle.aem.instance.tasks.InstanceRcp
 import com.cognifide.gradle.aem.instance.tasks.InstanceSatisfy
 import com.cognifide.gradle.aem.instance.tasks.InstanceSetup
 import org.gradle.testfixtures.ProjectBuilder
