+++ src/test/resources/com/cognifide/gradle/aem/common/instance/tail/10-logs-error.log
