+++ src/test/resources/com/cognifide/gradle/aem/common/instance/tail/aggregating/disjoint/first-chunk-error.log
