+++ src/test/resources/com/cognifide/gradle/aem/common/instance/tail/aggregating/disjoint/second-chunk-error.log
