+++ src/test/resources/com/cognifide/gradle/aem/common/instance/tail/aggregating/multiline/incomplete-multiline-logs-error.log
