+++ src/test/resources/com/cognifide/gradle/aem/common/instance/tail/aggregating/multiline/multiline-long.log
