+++ src/test/resources/com/cognifide/gradle/aem/common/instance/tail/aggregating/overlapping/first-chunk-error.log
