+++ src/test/resources/com/cognifide/gradle/aem/common/instance/tail/aggregating/overlapping/second-chunk-error.log
